requirements.txt
django
tweepy
gunicorn
psycopg2
djangorestframework
whitenoise
dj-database-url